pointillist brush
This was an added challenge for campers during Creative Coding Week with Spark. The lesson intended to teach campers about the get() function in P5.Image and encourage combining past projects together (In this case, it was the brush lesson, where campers learned about variables/functions relating to the mouse).