Design Your Own Character
This was a very simple demo for a "Design Your Own Character" Lesson in p5js. It taught students about the importance of order in written code, the difference in axes compared to a normal x/y grid, and how to draw shapes to make a character. This was updated for an "Animate Your Own Character" which taught the usage of variables in terms of position.