Pointillism + GIFs
This was a challenge I gave to one student and I wanted to implement it myself in Java. The program takes in the frames of the GIF (have to be separate, I used a GIF splitting website) and puts them into an array, and each frame is displayed to animate the GIF.  I overlaid the image with the same technique I used in other pointillism demos to alter the visual output. Loading in an already split gif was more of a job for Processing than p5, according to the camper who bailed on loading 20 images into the browser.